feat: Started Client Code
Moved modules around to make it easier to create local
jar in Maven repository so that we could easily import
use local jars in different modules. Moved alot of the server
files into the correct directory

Started the client code. Currently it is just a wrapper
that can easily GET/POST/PUT data into the server. Next up
is to start creating mechanism to easily create the report
of client and then creating the mechanism to read which
messages + what order to send.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>