feat: Add ability to parse operation chain
Coordinator client can now load preconfigured messages
and preconfigured operations which will allow us to
easily specify a chain of messages to send to the user.
The next step is to hook up everything together
for the coordinator client. This means that it will
be able to run an entire operation chain just by specifying it.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>