feat: Add ability to run an entire operations chain
The coordinator can run an entire operation chain appropriately.
Furthermore, it has the ability to stop an operation chain if
one operation fails. This is for the situation when you have
a POST request and a GET request right after. You wouldn't
want to fail a post and then run a GET right after.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>