refactor: Remove unnecessary class and add config options
Added two new fields into the OperationChain class which will
allow a user to specify whether they want all operations
in the chain to be run sequentially or in parallel. Furthermore,
they can specify that the operation should fail fast.

Removed unnecessary OperationChain in Operations as there was
one in models package.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>