feat: Add the ability to run operations in parallel
Added the ability to run operations in parallel. For implementation,
the future will wait for all operations to complete before returning
but it will only use the first failure. Therefore, you could have
multiple failures but only the first failure will be used.

Next step is to automatically generate the report.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>