feat: Improve Metrics Collector
Metrics Collector has access to more metrics and the ability
to store which messages weren't appropriately received.
It now creates a list of MismatchResponses which detail which
messages were incorrectly received and for what reason.

This should now easily allow us to add the ability to
create a report from scratch.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>