feat: Add ability to create string report
Moved metrics from collector into a class which handles
all general statistics. This has cleaned up the code significantly
so we only need to pass in a single object instead of 6 parameters.

Added a reporter class which has the ability to convert the
GeneralStatics and a List of MismatchedResponses into a fully
operation report. The next step is to output a file but currently
we can see the report in the log lines.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>