feat: Start creating terminal client
Created a terminal client which would allow a user to easily
interact with the client using the terminal. Added mockito
to easily test the input and output environment.

Started adding the ability to interact with the terminal.
Currently the terminal just has the ability to recognize
what inputs are in the environment but working on actually
creating the function to run the options.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>