feat: Complete Terminal Client
Client can now read from terminal and create report correctly.
Some refactoring was needed to be done because we needed
to execute some blocking operations which caused
the vertx to go crazy and spit a bunch of exceptions.
Refactored most of the methods to execute blocking
when needed (still need to do JSON object)

Also need to delete two files from repo so that
we have those folders.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>