FEAT: Add in Client (#4)
* feat: Started Client Code

Moved modules around to make it easier to create local
jar in Maven repository so that we could easily import
use local jars in different modules. Moved alot of the server
files into the correct directory

Started the client code. Currently it is just a wrapper
that can easily GET/POST/PUT data into the server. Next up
is to start creating mechanism to easily create the report
of client and then creating the mechanism to read which
messages + what order to send.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to read messages from file

Client now has the ability to read in messages from file which
will allow us to easily specify what messages to send to the
server. Next up is to make the mechanism which will
map ids to the expected message in the server.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to parse operation chain

Coordinator client can now load preconfigured messages
and preconfigured operations which will allow us to
easily specify a chain of messages to send to the user.
The next step is to hook up everything together
for the coordinator client. This means that it will
be able to run an entire operation chain just by specifying it.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Coordinate POST/GET/PUT methods

Coordinator can now send various HTTP requests and will automatically
keep track of what it expects which messages are associated
with which IDs.

Next step is to run an entire operations chain.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to run an entire operations chain

The coordinator can run an entire operation chain appropriately.
Furthermore, it has the ability to stop an operation chain if
one operation fails. This is for the situation when you have
a POST request and a GET request right after. You wouldn't
want to fail a post and then run a GET right after.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* refactor: Remove unnecessary class and add config options

Added two new fields into the OperationChain class which will
allow a user to specify whether they want all operations
in the chain to be run sequentially or in parallel. Furthermore,
they can specify that the operation should fail fast.

Removed unnecessary OperationChain in Operations as there was
one in models package.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add the ability to run operations in parallel

Added the ability to run operations in parallel. For implementation,
the future will wait for all operations to complete before returning
but it will only use the first failure. Therefore, you could have
multiple failures but only the first failure will be used.

Next step is to automatically generate the report.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Improve Metrics Collector

Metrics Collector has access to more metrics and the ability
to store which messages weren't appropriately received.
It now creates a list of MismatchResponses which detail which
messages were incorrectly received and for what reason.

This should now easily allow us to add the ability to
create a report from scratch.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to create string report

Moved metrics from collector into a class which handles
all general statistics. This has cleaned up the code significantly
so we only need to pass in a single object instead of 6 parameters.

Added a reporter class which has the ability to convert the
GeneralStatics and a List of MismatchedResponses into a fully
operation report. The next step is to output a file but currently
we can see the report in the log lines.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to create specified Report file

Added the ability to asynchronously create the report file
in a specified location. Also added the various unit tests
needed.

Added checkstyle check to ensure that all parameters are
final to enable functional style.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Start creating terminal client

Created a terminal client which would allow a user to easily
interact with the client using the terminal. Added mockito
to easily test the input and output environment.

Started adding the ability to interact with the terminal.
Currently the terminal just has the ability to recognize
what inputs are in the environment but working on actually
creating the function to run the options.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Create run loop

Created run loop which will now successfully
handle multiple requests and create the report
when the user decides to exit.

Next step is to hook it up to main.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Complete Terminal Client

Client can now read from terminal and create report correctly.
Some refactoring was needed to be done because we needed
to execute some blocking operations which caused
the vertx to go crazy and spit a bunch of exceptions.
Refactored most of the methods to execute blocking
when needed (still need to do JSON object)

Also need to delete two files from repo so that
we have those folders.

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* chore: delete files and add missing test

Deleted generated files and added missing JSONobject blocking
test

Contributes to: #2

Signed-off-by: Julian Goh <juliangohsl@gmail.com>