feat: Generate JAR with dependencies
Added the ability for the project to create JARs for each submodule
with the appropriate dependencies built in. This will allow us to
have a single base image with the JAVA dependencies that we
can build on top of for each image.

The next step is to set up the appropriate environment for each
image by copying the relevant files and envars in the image.

Contributes to: #6

Signed-off-by: Julian Goh <juliangohsl@gmail.com>