feat: Create Docker images
Created base image and accompanying component docker images which should
allow us to dockerize these components to run on any system.

Made some minor changes to the server code to make it easy
to set up envs through a getOrDefault method

Next step is to create run script which should allow us to easily run
client and the server in terminals.

Contributes to: #6

Signed-off-by: Julian Goh <juliangohsl@gmail.com>