feat: Add interface for general distributed algorithm
Created a general interface which will allow a user to create
a distributed algorithm that is able to be loaded into the server.
This distributed algorithm will run its own verticle whilst inside
the server and the everytime the coordination handler receives  a
message, it will send a message to the verticle holding the algo
to consume a message.

According to the following source on medium (posted in the issue)
the verticle will spawn a thread to consume the message and if
there are too many messages being consumed at once, it will have
to wait.

Next step is to flush out the API and add the ability to dynamically
load in the JAR into the server.

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>