feat: Add dynamically loading of algorithm and activiation
Added the ability to now dynamically load generated jars
at run time into the the algorithm. The loaded jar
runs in another verticle and whenever the server receives
a coordinate message, this will trigger the server to run
the code so it runs on every message (vertx will
handle the thread management). A few things that are needed
to ensure that the JAR can be used is, the user needs to:
1. Implement Distributed Algorithm as a class and implement
   the ConsumeMessage() method. The user doesn't need
   a main class because the assumption is that the server
   will target this class and instatiate it at runtime.
2. The user needs to pass in both the JAR location
   and the package name of the class to instantitate

Note that we have a full integration style package
in the project in the 'test' module. This module
is a simple example of a Distributed Algorithm
that simply copies the message on the queue.
However, the tests for this module serve as a way to
validate both the api and dynamically loading the jar.

The next steps steps are to fix the Docker run script
and make sure that the unit tests works because it
currently relies on the Test module to have its
jar generated.

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>