Feat: Add ability to load custom Distributed Algorithm (#13)
* feat: Add interface for general distributed algorithm

Created a general interface which will allow a user to create
a distributed algorithm that is able to be loaded into the server.
This distributed algorithm will run its own verticle whilst inside
the server and the everytime the coordination handler receives  a
message, it will send a message to the verticle holding the algo
to consume a message.

According to the following source on medium (posted in the issue)
the verticle will spawn a thread to consume the message and if
there are too many messages being consumed at once, it will have
to wait.

Next step is to flush out the API and add the ability to dynamically
load in the JAR into the server.

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add dynamically loading of algorithm and activiation

Added the ability to now dynamically load generated jars
at run time into the the algorithm. The loaded jar
runs in another verticle and whenever the server receives
a coordinate message, this will trigger the server to run
the code so it runs on every message (vertx will
handle the thread management). A few things that are needed
to ensure that the JAR can be used is, the user needs to:
1. Implement Distributed Algorithm as a class and implement
   the ConsumeMessage() method. The user doesn't need
   a main class because the assumption is that the server
   will target this class and instatiate it at runtime.
2. The user needs to pass in both the JAR location
   and the package name of the class to instantitate

Note that we have a full integration style package
in the project in the 'test' module. This module
is a simple example of a Distributed Algorithm
that simply copies the message on the queue.
However, the tests for this module serve as a way to
validate both the api and dynamically loading the jar.

The next steps steps are to fix the Docker run script
and make sure that the unit tests works because it
currently relies on the Test module to have its
jar generated.

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* fix: Server Docker container

Server Docker container is now fixed so that it works
appropriately with the loaded Distributed algorithm.
Just need to review code and then can merge!

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Update client script

Updated client script to make it fit the format of the server script.
This means it will print out the usage and you can configure the
SERVER_PORT to connect on

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>