Fix: Send Coordinate Messages On Post (#15)
* refactor: Separate Handler Packages

Moved the endpoint handlers into separate packages
that correspond to whether they are a coordination or
client endpoint.

Contributes to: #14

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Added the ability for post messages to be in queue

Added a new endpoint into the vertx which handles messages that
were posted in by the user. The Distributed Algorithm has a new
method which will specifically deal with new messages that
have been posted in. This will allow the user to specifically
format and deal with those messages specifically.

Contributes to: #14

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* chore: Add Log Lines

Added log lines to post

Contributes to: #14

Signed-off-by: Julian Goh <juliangohsl@gmail.com>