feat: Create Handler Gates
Created a Probabilistic Gate and an Unreachable Gate which
trigger depending on what the server status is.
Unreachable will prevent any messages from entering the
server and the Probabilistic Gate will fail
a message with probability x.

Currently only converted the POSTMessageHandler
but will now convert the rest of the handlers.
Also need to add in a new endpoin into the server
which will control the probabilistic failure rate.

Contributes to: #12

Signed-off-by: Julian Goh <juliangohsl@gmail.com>