feat: DELETE Message Handler
Finished the DELETE Message handler which will delete a message
based on a messageID. If the messageID is not present, it will
return an error

Also did some minor refactoring to move common functionality
in handlers to the abstract class.

Next step is to make sure that the PUT and DELETE messages
also get put on the coordination queue. Also to make sure
that the client has the functionality to delete messages.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>