refactor: Change to query params and add field
Changed all handlers to use query params (this will
cause errors in the client). This is done because it would
match what PUT handler is doing. Also changed PUT handler
to do what its supposed to do instead of just putting
a message at a specific id, it will now get the original
id message, delete that key and put it in the new id.

We still need to test to make sure that the DELETE,
POST, PUT requests get appropriately put on the
initial queue. We also need to change the client
API to adapt to the new query param model
and the PUT/DELETE handler.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>