Feat: Add DELETE Handler (#27)
* feat: DELETE Message Handler

Finished the DELETE Message handler which will delete a message
based on a messageID. If the messageID is not present, it will
return an error

Also did some minor refactoring to move common functionality
in handlers to the abstract class.

Next step is to make sure that the PUT and DELETE messages
also get put on the coordination queue. Also to make sure
that the client has the functionality to delete messages.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* refactor: Change to query params and add field

Changed all handlers to use query params (this will
cause errors in the client). This is done because it would
match what PUT handler is doing. Also changed PUT handler
to do what its supposed to do instead of just putting
a message at a specific id, it will now get the original
id message, delete that key and put it in the new id.

We still need to test to make sure that the DELETE,
POST, PUT requests get appropriately put on the
initial queue. We also need to change the client
API to adapt to the new query param model
and the PUT/DELETE handler.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Delete PUT Handler

Deleted the PUT Handler because we should simplify the model so
that we have to deal with less situations. Also the functionality
is still there as a user just needs to delete the original
key and put it into a new key.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add DELETE capability into Client

Added the ability to delete messages from the server in the
client and the metrics collector in the server.

The next step make sure that the DELETE adds a message into the
client queue.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add DELETE on message request

Added the ability DELETE Messages to appear
on  the initial message queue

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* test: Add missing test

Added missing metrics test

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>