FEAT: Client Expected Check (#28)
* feat: Add the ability to check expected status code

Checked the expected status code to see whether or not
it is an expected fail. This has only been added
to the DELETE handler so far.

Contributes to: #26

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Convert GET to Expected

GET now works with the expected model. Not going to do
POST because we will always assume that POST works.

Contributes to: #26

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add POST Expected

Managed to add expected check for POST and also modified
the other checks to expect the situation when a server
is unreachable.

Contributes to: #26

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* chore: Add Teardown

Added TearDown into the tests to make sure things are ok.

Contributes to: #26

Signed-off-by: Julian Goh <juliangohsl@gmail.com>