feat: LeadershipElection update leader
Added ability for the leadership election algorithm to
update the leader and follower configuration when
method is triggered. This is done without
sending labels to other servers because all servers
have access to the same config.

Next step is to add the broadcast method and
then add the unknown check to addCoordinateMessage.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>