feat: Add integration test for broadcast
Refactored tests so that we would be able to deploy
multiple servers (in the same verticle) in a test.
Had to do a lot of refactoring to make the
distributed verticle state as the format-address
was a static method and we needed to introduce state into the
method since we were deploying multiple servers. Otherwise,
it would be sending an update request to the same verticle
and would NPE as message would be empty.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>