feat: Integration test zookeeper leadership
Did some major refactoring of the distributed algorithm to allow
for us to introduce state into the distributed algorithm
(refactored configuration to be an object) so that
we could set up different states.

Also added getters and setters to the config object.
Added an integration test which will test that the servers
exchange messages.

Next step is to add log lines.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>