Feat: Create Zookeeper Leadership Election Algorithm (#37)
* feat: Initial Creation of Zookeeper Algorithm Project

Initial repo dump

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Candidate Information test

Created tests for the CandidateLeadershipInformation which will
deal with how the the servers exchange messages with one another.
Also started working on the broadcast handler but have not
completed the handler as of yet.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* test: Broadcast handler

Finished tests for broadcast handler.

Next step is to finish the leadership election handler

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: CandidateInformationRegistry

Created a registry which handles the information about
which servers have produced what candidate number.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: LeadershipElection update leader

Added ability for the leadership election algorithm to
update the leader and follower configuration when
method is triggered. This is done without
sending labels to other servers because all servers
have access to the same config.

Next step is to add the broadcast method and
then add the unknown check to addCoordinateMessage.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add integration test for broadcast

Refactored tests so that we would be able to deploy
multiple servers (in the same verticle) in a test.
Had to do a lot of refactoring to make the
distributed verticle state as the format-address
was a static method and we needed to introduce state into the
method since we were deploying multiple servers. Otherwise,
it would be sending an update request to the same verticle
and would NPE as message would be empty.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Integration test zookeeper leadership

Did some major refactoring of the distributed algorithm to allow
for us to introduce state into the distributed algorithm
(refactored configuration to be an object) so that
we could set up different states.

Also added getters and setters to the config object.
Added an integration test which will test that the servers
exchange messages.

Next step is to add log lines.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* chore: Logging and docs

Added logging and docs to make the algorithm easier to read.

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* docs: Minor fixes

change Main to main

Contributes to: #36

Signed-off-by: Julian Goh <juliangohsl@gmail.com>