feat: Add follower write to state
Added the ability for the follower to add proposal to
state and then when it receives the commit message,
it will perform the state update. This was done
by creating an async state update function in
state class. This still needs to be added to leaderWrite

Next step is to do the integration and add to leader write.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>