feat: Add integration test
Added full integration test which tests that a follower and
leaders get updated when the leader receives a write request.
Added additional functionality into the server which
allows servers to not process the request immediately
before putting it into the client queue.

Last step is to do the follower forwarding.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>