Feat: Zookeeper Write (#41)
* feat: Begin Write protocol

Started to creating the shared local history state which will
be used with the rest of the phases. Currently can add
and update state.

Next step is to work on the proposal and should delete the
tools package as it is useless.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add tracker

Added acknowledged and committed proposal tracker class
which will help the leader keep track of the inflight
proposals.

Next step is to actually implement the acknowledgement
and proposal phase.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Leader initial broadcast

Added the functionality which would the leader server to
broadcast the initial state update to the rest of the servers.
Also added the ability for the server to differentiate what
kind of message it is.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add followerWriteHandler

Added the follower write handler which can send the
acknowledgement and commit acknowledgement.

The next step is to add epoch checks.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add LeaderWrite to WriteHandler

Added LeaderWrite functionality to the writehandler.
A function that will automatically increment the counter
and broadcast initial proposal. Another function to add proposals
and if majority of proposals are acknowledged, then it will send
a commit message to servers.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add follower write to state

Added the ability for the follower to add proposal to
state and then when it receives the commit message,
it will perform the state update. This was done
by creating an async state update function in
state class. This still needs to be added to leaderWrite

Next step is to do the integration and add to leader write.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Test fixes and add state update to leader

Added state update to the leader and fixed
some tests.

Next step is to do the integration tests.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add integration test

Added full integration test which tests that a follower and
leaders get updated when the leader receives a write request.
Added additional functionality into the server which
allows servers to not process the request immediately
before putting it into the client queue.

Last step is to do the follower forwarding.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add follower forward

Follower will now Client Requests to the leader.

Contributes to: #34

Signed-off-by: Julian Goh <juliangohsl@gmail.com>