Feat: Discovery Phase (#42)
* fix: Test fixes

Add awaits to process state update and increased timeout.

Contributes to: #38

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Start DiscoveryHandler

Started discovery handler and fixed a bug where more than one write
requests would fail server propagation (due to not setting the
correct index).

Contributes to: #38

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Create Follower and Leader discovery handler

Created the follower handler which will broadcast the follower's
latest zxid. The leader handler will deal with collecting zxids
and setting up the next step to synchronize.

Contributes to: #38

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Complete DiscoveryHandler

Finished the DiscoveryHandler which will now be able to
easily handle the coordination message and broadcasting.

Contributes to: #38

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Added the discovery phase into the integration test

Added the discovery phase as part of our zookeeeper algorithm
and added it into our unit test. Also added the bit where
the leader will broadcast its new state to the followers
and the followers will update.

Contributes to: #38

Signed-off-by: Julian Goh <juliangohsl@gmail.com>