Feat: Dead Letter Queue (#47)
* feat: Add dead letter queue

Added dead letter queue to server so that we can retry
failed messages.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* fix: Address message out of order bug

Addressed bug where if messages were out of order then the
state wouldn't update correctly and wouldn't process any
messages. Done by keeping track of processed ids.

Also added addedID tracker which keeps track of added
ids to prevent duplicate states.

Contributes to: #46
Contributes to: #43

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* fix: fix issue with server config loading

Looking at wrong env var.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* fix: Fix client tests and add dead letter queue

Added the dead letter queue for the client requests and
cooridnation messages. Also performed various fixes for
the client code which was missing the async calls.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* refactor: Move zookeeper message handling to new class

Added the zookeeper message handling to a new class so that
we can process messages the same way when we use the deadletter
queue.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Fixed Retry Verticle

Instead of having the retry verticle retrying the original message
using message handler (which could lead to duplicate messages),
retry verticle will retry whatever messages failed to send. This
has meant changing the handler code to add failed coordination
messages when it fails

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Integrate retry verticle into server

Integrated the retry verticle into the server so now
the server will retry failed messages. Also realized
that we could get rid of the client dead letters
as they will have been processed to coordination
messages by the end.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add Server Stage

Added server stage into the algorithm so now we will be able
to tell what stage the server is currently in.

Next step is to implement the stage message check.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add stage mechanism

Added stage mechanisms which will now check for what
stage the message is in and process it accordingly.
If a message is from an earlier stage, it will be dropped
else it will be processed or put on the dead letter queue.

Contributes to: #46

Signed-off-by: Julian Goh <juliangohsl@gmail.com>