Verified Test Run (#49)
* feat: Add zookeeper set up script

Added the zookeeper set up script so now it
will automatically start a server. Also
fixed a small issue where wrong env was being
referenced which lead to not being able to load jar.

Contributes to: #48

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Docker compose file

Created a docker compose file which will allow us to kick off
multiple servers at once. We can also configure the number of servers
and the ports they are on.

Contributes to: #48

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat:  setup and tear down

Added setup and teardown scripts which will allow us to
kick off the servers and stop them when needed.

Have confirmed that we can conenct a client to the servers
and they replicate effectively.

Contributes to: #48

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* niggle: remove unneeded file

Contributes to: #48

Signed-off-by: Julian Goh <juliangohsl@gmail.com>