feat: Metrics Collector (#58)
* feat: Initial creation of metrics collector

Created the initial sub-directory for the metrics collector
and created the endpoints file as well as the test to make
sure it is correct.

Contributes to: #50

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add tracking object and entry creation

Created the object which will track the successes and fails
of each message. Also created th object which will create
the report object.

Contributes to: #50

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add tracking endpoint

Created a tracking endpoint which will track various statistics
for a given message. Also created the object that will keep
track of this.

Contributes to: #50

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add report endpoint

Created the report endpoint which will create the report. Uses
the report creator class to create the class via a filter.

Contributes to: #50

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Metrics collector docker image

Created the metrics collector docker image which will allow us
to start the docker image at a specific host and port.
Finished up the metrics-collector jar so that it can be pulled into the
docker image. Also updated the zookeeper kickoff script to setup
metrics collector.

Contributes to: #50

Signed-off-by: Julian Goh <juliangohsl@gmail.com>