fix: Various fixes (#62)
* fix: Various fixes

Implemented various fixes in the zookeeper algorithm and the
metrics collector. Also fixed the error with the logger
and ensured that we can perform a test run of zookeeper on
live servers. Now have verified that it works and generates
the report.

Contributes to: #55

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* test: Add start up time out

Added start up time out to ensure that we can start the
server correctly in the tests

Contributes to: #55

Signed-off-by: Julian Goh <juliangohsl@gmail.com>