Gossip Write (#68)
* feat: Start gossip protocol module

Created the initial module and added some basic classes.
Added the models that will be necessary for future operations
as well as the state class.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add write handler

Added the write handler and implemented the structure of the
tests that will be needed for later.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Implemented reply handler

Implemented a reply handler which will reply to the servers
based on if it has seen the message. Also implemented the
messageHandler which will decide which path to run through
based on type of message.

Next step is to implement the retry verticle and integrate it.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Complete Retry verticle

Added the retry verticle and updated the handlers so that
they will add to the dead letter queue on fail.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Complete Integration of Algorithm

Algorithm is now ready to be deployed in an integration test
and tested for functionality.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: implement integration tests

Implemented the integration tests and also added the ability
to prevent infinite transactions. This is done by tracking
the keys which have become inactive.

Contributes to: #63

Signed-off-by: Julian Goh <juliangohsl@gmail.com>