Feat: Synchronize Handler (#70)
* feat: Added the ability to communicate state

Added the ability for the servers to communicate state through
the synchronization hanlder. Created the corresponding objects
to make it easier to handle this stuff.

Contributes to: #65

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to synchronize state

Added the synchronize state ability and integrated it into
the metrics handler. Next step is to add an ocassional
synchronize broadcast into the retry verticle.

Contributes to: #65

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add synchronize verticle

Added the synchronize verticle.

Contributes to: #65

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: add test-run files

Added kickoff and teardown files.

Contributes to: #65

Signed-off-by: Julian Goh <juliangohsl@gmail.com>